feat(nodes): load ZNOD (zlib-compressed extended nodes) + fix >32767-sidedef maps
- ZNOD extended nodes: inflate the zlib body (isolated in w_inflate.c because
  <zlib.h> clashes with p_spec.h's 'close' enum) then run the existing XNOD
  parser -- ZNOD is just deflated XNOD. Link -lz. Verified end-to-end on a
  ZNOD-recompressed LoR MAP13 (identical vert/seg/node counts to XNOD).
- Limit-removing fix: P_LoadLineDefs read sidenum as a SIGNED short, so maps
  with >32767 sidedefs (LoR MAP13 has 60423) got a negative index that slipped
  past the -1 test and indexed sides[] out of bounds -> NULL frontsector, crash
  in P_GroupLines. Read sidenum unsigned (0xFFFF = none) and widen line_t.sidenum
  to int.

Follow-up: MAP13 still needs a blockmap rebuild (its 76134-short BLOCKMAP
exceeds the 16-bit offset limit) -- separate commit.

+++ files/r_defs.h
@@ -212,7 +212,9 @@ typedef struct line_s
 
     // Visual appearance: SideDefs.
     //  sidenum[1] will be -1 if one sided
-    short	sidenum[2];			
+    // int (not short): limit-removing maps can have >32767 sidedefs, which overflow a short
+    // index (Legacy of Rust MAP13 has 60423) -- P_LoadLineDefs reads them unsigned into this.
+    int		sidenum[2];
 
     // Neat. Another bounding box, for the extent
     //  of the LineDef.
